Move module package
