+++ android/src/main/java/nabidreams/audiorecorder/AudioRecorderModule.kt
@@ -1,4 +1,4 @@
-package nabidreams
+package nabidreams.audiorecorder
 
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
