+++ android/src/main/java/nabidreams/audiorecorder/AudioRecorderPackage.kt
@@ -1,13 +1,11 @@
-package nabidreams
+package nabidreams.audiorecorder
 
 import java.util.Arrays
-import java.util.Collections
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
-import com.facebook.react.bridge.JavaScriptModule
 
 class AudioRecorderPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
